backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
pgvector==0.3.1
psycopg2-binary==2.9.9
boto3==1.34.131
botocore==1.34.131
PyMuPDF==1.24.5
pdfplumber==0.11.1
pytesseract==0.3.10
Pillow==10.3.0
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.8
langchain-anthropic==0.1.15
tiktoken==0.7.0
openai==1.35.3
anthropic==0.29.0
numpy==1.26.4
pandas==2.2.2
pyarrow==16.1.0
tqdm==4.66.4
aiofiles==23.2.1
httpx==0.27.0
redis==5.0.6
hiredis==2.3.2
structlog==24.2.0
razorpay==1.4.1
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
pyathena==3.10.0
boto3==1.34.131
